+++ connectivity_features/gen_cross_spectrum.m
@@ -16,7 +16,7 @@
 % John M. O' Toole, University College Cork
 % Started: 17-03-2017
 %
-% last update: Time-stamp: <2017-03-28 17:21:02 (otoolej)>
+% last update: Time-stamp: <2017-03-29 10:06:37 (otoolej)>
 %-------------------------------------------------------------------------------
 function [pxy,Nfreq,f_scale,fp]=gen_cross_spectrum(x,y,Fs,param_st)
 
@@ -27,65 +27,41 @@ freq_bands=param_st.freq_bands;
 spec_method=param_st.method;
     
     
-
 % remove NaNs:
 x(isnan(x))=[];
+
+
+%---------------------------------------------------------------------
+% Welch cross-PSD 
+%---------------------------------------------------------------------
+[S_x,Nfreq,f_scale,win_epoch]=gen_STFT(x,L_window,window_type,overlap,Fs,1);
+[S_y,Nfreq,f_scale,win_epoch]=gen_STFT(y,L_window,window_type,overlap,Fs,1);
+
+S_xy=S_x.*conj(S_y);
 
 
 switch lower(spec_method)
   case 'psd'
     %---------------------------------------------------------------------
-    % Welch cross-PSD 
+    % mean for Welch PSD
     %---------------------------------------------------------------------
-    [S_x,Nfreq,f_scale,win_epoch]=gen_STFT(x,L_window,window_type,overlap,Fs,1);
-    [S_y,Nfreq,f_scale,win_epoch]=gen_STFT(y,L_window,window_type,overlap,Fs,1);
-
-    S_xy=S_x.*conj(S_y);
-
-    pxy=nanmean(S_xy)';
-    N=length(pxy);
-
-    % normalise (so is similar to Welch's PSD):
-    E_win=sum(abs(win_epoch).^2)./Nfreq;
-    pxy=(pxy./(Nfreq*E_win*Fs));
-    
+    pxy=nanmean(S_xy)';    
     
   case 'robust-psd'
     %---------------------------------------------------------------------
-    % similar to Welch but use median instead of mean value
+    % median for robust PSD
     %---------------------------------------------------------------------
-    [S_x,Nfreq,f_scale,win_epoch]=gen_STFT(x,L_window,window_type,overlap,Fs,1);
-    [S_y,Nfreq,f_scale,win_epoch]=gen_STFT(y,L_window,window_type,overlap,Fs,1);
-
-    S_xy=S_x.*conj(S_y);
-
     pxy=nanmedian(S_xy)';
-    N=length(pxy);
-    
-    % normalise (so is similar to Welch's PSD):
-    E_win=sum(abs(win_epoch).^2)./Nfreq;
-    pxy=(pxy./(Nfreq*E_win*Fs));
-
-    
-  case 'periodogram'
-    %---------------------------------------------------------------------
-    % Periodogram
-    %---------------------------------------------------------------------
-    X=fft(x);
-    Y=fft(y);    
-    pxy=X.*conj(Y);
-
-    % +ve frequencies only:
-    N=length(pxy); Nh=floor(N/2); Nfreq=N;
-    pxy=pxy(1:Nh+1)';
-
-    pxy=pxy./(Fs*N);
-    
     
   otherwise
     fprintf('unknown spectral method ''%s''; check spelling\n',spec_method);
     pxy=NaN; f_scale=NaN; fp=NaN;
 end
+
+
+% normalise (so is similar to Welch's PSD):
+E_win=sum(abs(win_epoch).^2)./Nfreq;
+pxy=(pxy./(Nfreq*E_win*Fs));
 
 
 N=length(pxy);