+++ neural_parameters.m
@@ -131,22 +131,19 @@ feat_params_st.rEEG.freq_bands=FREQ_BANDS;
 %---------------------------------------------------------------------
 % C. connectivity features
 %---------------------------------------------------------------------
-% how to estimate the spectrum for 'spectral_flatness', 'spectral_entropy',
-% spectral_edge_frequency features:
+% how to estimate the cross spectrum for the coherence function:
 % 1) PSD: estimate power spectral density (e.g. Welch periodgram)
 % 2) robust-PSD: median (instead of mean) of spectrogram 
-% 3) periodogram: magnitude of the discrete Fourier transform
-feat_params_st.connectivity.method='periodogram'; 
+feat_params_st.connectivity.method='PSD'; 
 
-% feat_params_st.connectivity.freq_bands=[FREQ_BANDS(1) FREQ_BANDS(end)];
 feat_params_st.connectivity.freq_bands=FREQ_BANDS;
-feat_params_st.connectivity.L_window=3.12; % PSD window in seconds
-feat_params_st.connectivity.overlap=21; % PSD window percentage overlap
+feat_params_st.connectivity.L_window=2; % PSD window in seconds
+feat_params_st.connectivity.overlap=50; % PSD window percentage overlap
 feat_params_st.connectivity.window_type='hamm'; % PSD window type
 % find lower coherence limit using surrogate data?
 % (number of iterations required to generate null-hypothesis distribution;
 %  set to 0 to turn off)
-feat_params_st.connectivity.coherence_surr_data=500; 
+feat_params_st.connectivity.coherence_surr_data=100; 
 % alpha value for null-hypothesis disribution cut-off:
 feat_params_st.connectivity.coherence_surr_alpha=0.05;